good styling, integration. closing this branch because the next one will be experimental
