+++ Python Rebuild/chs-source/src/Components/venueReact/chsPourHouse/chsPourHouse.css
@@ -1,0 +1,17 @@
+
+
+table {
+  border: 1px solid black;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: space-around;
+}
+
+th {
+  border: 1px solid black;
+}
+
+td {
+  border: 1px solid black;
+}
