+++ Python Rebuild/chs-source/src/Components/venueReact/musicFarm/musicFarm.css
@@ -1,0 +1,20 @@
+
+div {
+
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: space-around;
+}
+
+table {
+  border: 1px solid black;
+}
+
+th {
+  border: 1px solid black;
+}
+
+td {
+  border: 1px solid black;
+}
